Static Array enum dispatch (#578)
This PR is an experiment into making arrays using static dispatch.

The enum_dispatch crate breaks IntelliJ's code analysis, so for now I've
written some macros that make it somewhat reasonable to dispatch the
ArrayEncoding trait onto each of the encodings themselves. We can switch
to this crate later with minimal code changes if/when IntelliJ is fixed.

An Array is an enum of the possible enc encodings. Both the Array enum
and all encoding arrays implement the ArrayEncoding trait that holds
common methods.

I've hacked together a broken implementation of the "add" compute
function that demonstrates the niceties of matching over the lhs/rhs
encodings.

We could also implement this same approach for scalars.