IPC Terminator (#267)
This PR adds an end-of-stream marker to the IPC format such that the
format itself can be embedded within other streams without consuming any
further bytes.

It also fixes a bug whereby an ArrayStreamReader would return an error
if it saw anything but a ChunkMessage, which would occur whenever we
wrote multiple arrays into a single stream. There is now a
StreamMessageReader that essentially allows peeking at the next message
in order to check if it should be processed.