fix: canonicalization of chunked ExtensionArray (#499)
Since #346, we've been canonicalizing chunked ExtensionArray's
incorrectly. Rather than unwrapping each chunk to its backing storage
array we've just been building a new chunkedarray of the ExtensionArray
chunks.

Before #480 , this was actually causing DateTimePartsCompressor to fail
in can_compress, so we weren't going down the compression codepath.

## The Fix

Fix `into_canonical` so that when we encounter a
`ChunkedArray(ExtensionArray(storage))` to yield an
`ExtensionArray(ChunkedArray(storage))` where each chunk is
canonicalized first (e.g. if you have chunks of DateTimePartsArray you
will end up with chunks of ExtensionArray(i64)).

Fix the DateTimePartsCompressor to canonicalize its input, to handle
that case where it may be a ChunkedArray.