smoketest for SamplingCompressor, fix bug in varbin stats
There are a number of failures in the past few months that only show up in an ETE test
like random_access benchmark. However, random_access benchmark is very slow to run in
CI.

I'm adding a very tiny SamplingCompressor test that runs at CI time. It takes <1s
on my macbook, and uses 65k rows of synthetic data to try and get the compressor
to go down a few common paths.

It's not perfect, but it does trigger Constant, DateTimeParts, Dict and Bitpacking