Smoketest for SamplingCompressor, fix bug in varbin stats (#510)
There are a number of failures in the past few months that only show up
in an ETE test like random_access benchmark. However, random_access
benchmark is very slow to run in CI.

I'm adding a very tiny SamplingCompressor test that runs at CI time. It
takes <1s on my macbook, and uses 65k rows of synthetic data to try and
get the compressor to go down a few common paths.

It's not perfect, but it does trigger Constant, DateTimeParts, Dict and
could trigger Bitpacking if we implement #509.

Bonus: after writing this test and examining the compressed output in a
few scenarios, I was noticing that Dict compression was never kicking in
for out-of-order arrays. It looks like we were computing the
is_strict_sorted stat incorrectly for varbin.