Teach RunEnd take to respect its own validity (#691)
Before this change, the validity is completely ignored. The test added
in this PR fails in the develop branch. This change also takes the
desired indices from the validity and converts from an array of Booleans
to an array of set indices. This array of indices can be used as the
indices of a sprase array with the fill values as null.

Resolves #609.