Primitive Iterator API (#689)
This PR includes a new batched iterator API, that yeilds a `Vec<T>` and
validity data for batches of items from arrays. It can dynamically
dispatch over the type of underlying arrays, so recursive compression
should hold.
The PR includes implementations for primitve arrays, as well as ALP as a
test case for more complex encodings and constant array as its fairly
trivial.

I also added a bunch of benchmarks to various pieces it touches to have
some initial performance numbers and test basic correctness.