Sphinx docs skeleton
The Python ecosystem uses either Sphinx or MkDocs. The Astral folks
use MkDocs, but Arrow and I suspect the majority of the Python ecosystem
use Sphinx.

The most noticeable difference between the two is that MkDocs uses Markdown
and Sphinx uses reStructuredText. I have used Sphinx too long
to be bothered by rST but, IMO, a lot of people dislike
the syntax at first.

Translating between rST and Markdown will be annoying so we should
make a decision before too long. For now, I got the site started
with the tool I knew, Sphinx, so that is what I have added here.