FSSTCompressor (#664)
* Implements a new codec, FSSTCompressor using latest version of fsst-rs
library
* Adds new `metadata` field on CompressionTree to allow reuse between
the sampling and compressing stages. For example, we can save the ALP
exponents to not have to calculate them twice. This is very important
for FSST so that we save the overhead of training the table twice
* Adds new compression benchmark using the `lineitem` table's
`l_comment` column, with scalefactor=1, which is just over 6million
rows. By default this is loaded as a ChunkedArray with 733 partitions.
Compressing with FSST enabled takes 1.6s. Compressing on the
canonicalized array takes ~550ms. We should be able to speed this up by
at least ~2x, see
https://github.com/spiraldb/vortex/pull/664#issuecomment-2327213583, and
we can potentially do even better. We probably want to be able to FSST
compress a ChunkedArray directly so that we avoid the overhead of
training/compressing each chunk from scratch.