ALP compressor is better at roundtripping values (#736)
For those wondering about the details of the change. `fast_round` as
described in ALP paper still can leave the number as a float (found
empirically by the fuzzer with the values committed in the test in this
pr) and it's important to compare the decoded value after it has been
converted to int and not just rounded.