Fix `rye build`
`rye build` failed when it tried to build the sdist to produce a wheel. This failed because some
_transitive_ workspace dependencies (e.g. vortex-proto) were not present in the sdist.

The root cause appears to be a bug in maturin. I saw [a similar issue on their
repo](https://github.com/PyO3/maturin/issues/2117#issuecomment-2330245151). I shared a branch with
them that reproduces the behavior.

Until this issue is resolved, we can work around it by explicitly depending on every workspace
crate.

Two unrelated issues are also fixed here:

1. We must propagate the Rust toolchain to the sdist build environment.

2. `vortex-bytebool` was misnamed in the root Cargo.toml. I'm not sure why this isn't a problem for
other crates?