Canonicalize FSST arrays without branching (#785)
FSST saves the compressed bytes into a giant packed string heap (which
is stored in a `VarBinArray`).

Canonicalizing currently requires decompressing each string
one-at-a-time, so that we can get the uncompressed length to construct
the new canonical `VarBinAray`. It turns out that
`VarBinArray::bytes_at()` is incredibly expensive when it's backed by an
`ArrayView`.

This PR instead

* Decompresses the entire packed heap in a single call to the
decompressor. This lets us spend all of our time in the hot decompress
loop without expensive branches
* Stores the uncompressed lengths for each string. When canonicalizing,
we can turn this into an offsets array. Later when we are using
`VarBinView`, we can use the lengths directly to build our `BinaryView`
values.