Introduce MaybeCompareFn trait to allow for partial compare specializations (#768)
Currently an array needs to statically decide if it supports a compute
function, prior to seeing the RHS or other arguments.

This change allows the array's compute function to see the RHS, and then
decide it doesn't support the operation, enabling easier fallback.

This is only implemented for compare.

It has also highlighted the need for storage DType information to be
available to an ExtScalar. Either by adding it to ExtDType, or adding it
directly into that scalar variant. @a10y keen for your thoughts

Notes by @AdamGS:
To follow up @gatesn's work here, I added some other things to this PR:
- I cleared up the boolean compare implementation as they were basically
the arrow-based fallback we already had
- Added an ALP `MaybeCompareFn` implementation that shows how the new
interface looks

---------

Co-authored-by: Adam Gutglick <adam@spiraldb.com>