feat: ArrayView::child will throw if encoding not found
Array::child returns an Option<Array>. The only time it should
really ever return None is when you are passing the result to
ValidityMetadata::to_validity.

Currently, if an array child is encoded with an unknown encoding ID
(e.g. b/c it's not been linked into your binary) then you end up silently
failing, and swalloing the reason why the None was returned.

IMO this is an irrecoverable error and is one of the times it's better
to fail with a helpful message rather than forge on ahead.