feat: teach DeltaArray slice and scalar_at (#927)
More subtle than I expected.

DeltaArray is a sequence of chunks. All chunks except the last must be
"full", i.e. containing 1,024 values. The last chunk may contain as few
as one value and is encoded differently from the rest.

In this PR, I introduced an "offset" and a "limit". Together they enable
logical/lazy slicing while preserving full chunks for later
decompression. The offset is a value, less than 1024, which offsets into
the first chunk. The limit is less than or equal to 1024.