feat: teach FSSTArray to compress the offsets of its codes (#952)
The codes of an FSSTArray are a vector of binary-strings of one byte
codes or an escape code followed by a data. The offsets, unexpectedly,
grow quite large, increasing the file size (for example, the TPC-H
l_comment column with this PR is 78% the byte size of itself on
`develop`). Delta encoding notably decreases the size but also inflates
the compression time, seemingly proportional to the space savings (TPC-H
l_comment compresses in 111% of the time on `develop`).

---------

Co-authored-by: Robert Kruszewski <github@robertk.io>