feat: cost of for is ~0 (#967)
We fairly often perform FoR and then do not perform bitpacking because
we have exceeded the depth. FoR is very cheap and only worthwhile in
combination with bitpacking anyway, so setting its cost to zero should
improve compression ratio with a minor decrease in compresssion speed.

I've been poking around trying to make the compressor reliably faster
and along the way I noticed that this noticeably improves compression
ratio with a minor impact on compression speed.