feat: proto matches serde: f16 serializes as an unsigned integer (#992)
Protobufs does not support u16, so we extend to u32. Protobufs use a
[variable width integer
encoding](https://protobuf.dev/programming-guides/encoding/#varints) so
these should be encoded in at most three bytes.

I also flipped the round_trip test so that the expected value is on the
left and the actual value is on the right which, for reasons I cannot
explain, feels "right".