feat: marginally faster From<Vec<T>> for Scalar
When constructing stats, we convert into Scalars, which I suspect is unnecessary (given the element
dtype, we should know the dtypes of all the stats). Nevertheless, as long as we are constructing
Scalars, we should do it quickly!

At the cost of more macros and generated code, this reduced Bimbo compression runtime by 500ms (of
18s, so ~3%).