feat: concise benchmark names
Currently the Taxi and l_comment datasets are their own Criterion "group" which
causes them to be plotted separately. With this PR, there is one plot for
"compress time" which is measured in seconds and has every dataset.

I also eliminated the comparison to uncompressed Parquet because (except for
a tiny 1024 byte dataset) we always beat uncompressed Parquet. Parquet with
ZSTD is a sufficient comparison.

Finally, I removed some characters from several benchmark names in hopes that
the Markdown tables on GitHub are more legible.